+++ arviz/plots/backends/bokeh/khatplot.py
@@ -21,10 +21,11 @@ def plot_khat(
     xdata,
     khats,
     kwargs,
-    annotate,
+    threshold,
     coord_labels,
+    show_hlines,
     show_bins,
-    hlines_kwargs,  # pylint: disable=unused-argument
+    hlines_kwargs,
     xlabels,  # pylint: disable=unused-argument
     legend,  # pylint: disable=unused-argument
     color,
@@ -48,6 +49,10 @@ def plot_khat(
     }
 
     (figsize, *_, line_width, _) = _scale_fig_size(figsize, textsize)
+
+    if hlines_kwargs is None:
+        hlines_kwargs = {}
+    hlines_kwargs.setdefault("hlines", [0, 0.5, 0.7, 1])
 
     cmap = None
     if isinstance(color, str):
@@ -103,21 +108,21 @@ def plot_khat(
             fill_alpha=alphas,
         )
 
-    if annotate:
-        idxs = xdata[khats > 1]
+    if threshold is not None:
+        idxs = xdata[khats > threshold]
         for idx in idxs:
             ax.text(x=[idx], y=[khats[idx]], text=[coord_labels[idx]])
 
-    for hline in [0, 0.5, 0.7, 1]:
-        _hline = Span(
-            location=hline,
-            dimension="width",
-            line_color="grey",
-            line_width=line_width,
-            line_dash="dashed",
-        )
-
-        ax.renderers.append(_hline)
+    if show_hlines:
+        for hline in hlines_kwargs.pop("hlines"):
+            _hline = Span(
+                location=hline,
+                dimension="width",
+                line_color="grey",
+                line_width=line_width,
+                line_dash="dashed",
+            )
+            ax.renderers.append(_hline)
 
     ymin = min(khats)
     ymax = max(khats)
@@ -134,6 +139,7 @@ def plot_khat(
                 text=[bin_format.format(count, count / n_data_points * 100)],
             )
         ax.x_range._property_values["end"] = xmax + 1  # pylint: disable=protected-access
+
     ax.xaxis.axis_label = "Data Point"
     ax.yaxis.axis_label = "Shape parameter k"
 
@@ -141,7 +147,7 @@ def plot_khat(
         ax.y_range._property_values["start"] = -0.02  # pylint: disable=protected-access
     if ymax < 1:
         ax.y_range._property_values["end"] = 1.02  # pylint: disable=protected-access
-    elif ymax > 1 & annotate:
+    elif ymax > 1 & threshold:
         ax.y_range._property_values["end"] = 1.1 * ymax  # pylint: disable=protected-access
 
     show_layout(ax, show)