+++ CHANGELOG.md
@@ -33,7 +33,9 @@
 * `plot_elpd`, avoid modifying the input dict ([1477](https://github.com/arviz-devs/arviz/issues/1477))
 * Do not plot divergences in `plot_trace` when `kind=rank_vlines` or `kind=rank_bars` ([1476](https://github.com/arviz-devs/arviz/issues/1476)) 
 
+
 ### Deprecation
+* `plot_khat` deprecate `annotate` argument in favor of `threshold`. The new argument accepts floats ([1478](https://github.com/arviz-devs/arviz/issues/1478))
 
 ### Documentation
 * Reorganize documentation and change sphinx theme ([1406](https://github.com/arviz-devs/arviz/pull/1406))